Ansatz function for inclusive jets into a separate file. Started bringing systematics.C to modern times.
